Refactoring design
