+++ src/integrationTest/java/uk/gov/hmcts/reform/userprofileapi/integration/CreateNewUserProfileTest.java
@@ -141,5 +141,4 @@ public class CreateNewUserProfileTest {
 
     }
 
-
 }
