+++ src/main/java/uk/gov/hmcts/reform/userprofileapi/domain/service/ResourceCreator.java
@@ -1,9 +1,10 @@
 package uk.gov.hmcts.reform.userprofileapi.domain.service;
 
-import uk.gov.hmcts.reform.userprofileapi.infrastructure.clients.CreateUserProfileData;
+import uk.gov.hmcts.reform.userprofileapi.domain.entities.UserProfile;
+import uk.gov.hmcts.reform.userprofileapi.infrastructure.clients.RequestData;
 
-public interface ResourceCreator {
+public interface ResourceCreator<T extends RequestData> {
 
-    void create(CreateUserProfileData profileData);
+    UserProfile create(T profileData);
 
 }
