+++ src/main/java/uk/gov/hmcts/reform/userprofileapi/infrastructure/repository/UserProfileRetriever.java
@@ -1,8 +1,9 @@
-package uk.gov.hmcts.reform.userprofileapi.domain.service;
+package uk.gov.hmcts.reform.userprofileapi.infrastructure.repository;
 
 import org.springframework.dao.DataRetrievalFailureException;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.userprofileapi.domain.entities.UserProfile;
+import uk.gov.hmcts.reform.userprofileapi.domain.service.ResourceRetriever;
 import uk.gov.hmcts.reform.userprofileapi.infrastructure.clients.UserProfileIdentifier;
 
 @Service
@@ -19,7 +20,7 @@ public class UserProfileRetriever implements ResourceRetriever {
 
         String failureMessage = "Could not find resource from database with given identifier: ";
 
-        return querySupplier.getIdQuery(identifier).get()
+        return querySupplier.getQueryByIdentifier(identifier).get()
             .orElseThrow(() ->
                 new DataRetrievalFailureException(failureMessage + identifier.getValue()));
     }
