+++ src/main/resources/application.yaml
@@ -48,6 +48,7 @@ security:
     - "/v2/api-docs"
     - "/webjars/springfox-swagger-ui/**"
   authorisedServices:
+    - "ccd_gw"
 
 ### dependency configuration
 auth.idam.client.baseUrl: ${IDAM_URL:http://127.0.0.1:4501}
